requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Device control
phue2>=0.0.3              # Hue bridge control (Python 3.10+)
soco>=0.30.0              # Sonos control via UPnP

# TTS
edge-tts>=6.0.0           # Free Microsoft TTS (primary)
gtts>=2.5.0               # Google TTS (fallback)

# Database
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0

# Config
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP
httpx>=0.27.0             # Async HTTP client for ESPN API

# PC activity detection
psutil>=5.9.0             # Process monitoring for activity detection
pyaudio>=0.2.14           # Blue Yeti mic ambient noise detection

# Utilities
aiofiles>=23.0.0          # Async file operations
